Add UDP support for posting events.
This commit also changes the emitter client so that you can easily switch
between websockets and UDP as desired. UDP is faster, but does not detect when
events are not sent successfully.
