+++ lib/cube/collector.js
@@ -21,6 +21,9 @@ exports.register = function(db, endpoints) {
     endpoint("POST", "/1.0/event/put", poster),
     endpoint("POST", "/collectd", require("./collectd").putter(putter))
   );
+
+  //
+  endpoints.udp = putter;
 };
 
 function post(putter) {
