+++ lib/cube/server.js
@@ -1,6 +1,7 @@
 var util = require("util"),
     url = require("url"),
     http = require("http"),
+    dgram = require("dgram"),
     websocket = require("websocket"),
     websprocket = require("websocket-server"),
     static = require("node-static"),
@@ -157,8 +158,20 @@ module.exports = function(options) {
       meta = require("./event").putter(db);
       util.log("starting http server on port " + options["http-port"]);
       primary.listen(options["http-port"]);
+      if (endpoints.udp) {
+        util.log("starting udp server on port " + options["udp-port"]);
+        var udp = dgram.createSocket("udp4");
+        udp.on("message", function(message) {
+          endpoints.udp(JSON.parse(message.toString("utf8")), ignore);
+        });
+        udp.bind(options["udp-port"]);
+      }
     }
   };
 
   return server;
 };
+
+function ignore() {
+  // Responses for UDP are ignored; there's nowhere for them to go!
+}
